% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.R
\name{auto_corr}
\alias{auto_corr}
\title{Compute Autocorrelation}
\usage{
auto_corr(x, tau)
}
\arguments{
\item{x}{vector of distinct options of numbers or characters}

\item{tau}{numeric indicating the gap between subsequent numbers for which
the autocorrelation will be computed}
}
\value{
autocorrelation of \code{x} for the gap between subsequent numbers \code{tau}
}
\description{
Compute Autocorrelation
}
\details{
This function takes a vector \code{x} and uses it to compute the autocorrelation
for a specified gap tau \code{tau}. This is done by correlating the original vector \code{x}
with the same sequence starting \code{tau} values later. A high autocorrelation
for a specified value \code{tau} indicates that there is a high correlation
between numbers in a sequence that are \code{tau} values apart from each other.
}
\examples{
max_auto_corr(c(1, 1, 1, 2, 2, 2, 1, 2, 1, 2, 2, 2), tau = 2)
max_auto_corr(c(1, 1, 1, 2, 2, 2, 1, 2, 1, 2, 2, 2), tau = 3)

}
