% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{convert_to_matrix}
\alias{convert_to_matrix}
\title{helper function to transform a given vector to a matrix of first order
dependencies, specifying the frequency of all possible 2-digit long sequences
#' @param x vector of random numbers}
\usage{
convert_to_matrix(x, possible_responses)
}
\arguments{
\item{possible_responses}{number of available options in sequence}
}
\description{
helper function to transform a given vector to a matrix of first order
dependencies, specifying the frequency of all possible 2-digit long sequences
#' @param x vector of random numbers
}
\keyword{internal}
