% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularity_indices.R
\name{coupon_score}
\alias{coupon_score}
\title{Compute Coupon Score (see Ginsburg & Karpiuk)}
\usage{
coupon_score(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
Coupon Score of \code{x}
}
\description{
Compute average digit length required for all responses to occur
}
