% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{digram_rep}
\alias{digram_rep}
\title{Compute digram repetitions (see Ginsburg & Karpiuk, 1994)}
\usage{
digram_rep(x, possible_responses)
}
\arguments{
\item{x}{vector of random numbers}

\item{possible_responses}{number of available options in sequence}
}
\description{
Compute digram repetitions (see Ginsburg & Karpiuk, 1994)
}
