% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evans1978.R
\docType{data}
\name{evans1978}
\alias{evans1978}
\title{Two sequences of random numbers of length 100 generated by human participants.}
\format{
A data frame with 100 rows and 2 columns:
\describe{
  \item{sequence_one} {first sequence of random numbers}
  \item{sequence_two} {second sequence of random numbers}
}
}
\source{
Taken from Evans (1978; see reference below),
\url{https://doi.org/10.3758/BF03329617}.
}
\usage{
evans1978
}
\description{
Two sequences of random numbers of length 100 generated by human participants.
}
\references{
Evans, F.J. Monitoring attention deployment by random number generation:
An index to measure subjective randomness.
Bull. Psychon. Soc. 12, 35–38 (1978).
\url{https://doi.org/10.3758/BF03329617}
}
\keyword{datasets}
