% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginsburg1994.R
\docType{data}
\name{ginsburg1994}
\alias{ginsburg1994}
\title{One sequence of random numbers of length 100 generated by a human participant.}
\format{
A 100-digit-long vector of random numbers:
\describe{
  \item{ginsburg1994} {sequence of random numbers}
}
}
\source{
Taken from Ginsburg & Karpiuk, 1994; see reference below),
\url{doi:10.2466/pms.1994.79.3.1059}.
}
\usage{
ginsburg1994
}
\description{
One sequence of random numbers of length 100 generated by a human participant.
}
\references{
Ginsburg N, Karpiuk P. Random Generation: Analysis of the Responses.
Perceptual and Motor Skills. 1994;79(3):1059-1067.
\url{doi:10.2466/pms.1994.79.3.1059}
}
\keyword{datasets}
