% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.R
\name{max_auto_corr}
\alias{max_auto_corr}
\title{Compute the Maximum Autocorrelation}
\usage{
max_auto_corr(x, min_tau, max_tau)
}
\arguments{
\item{x}{vector of distinct options of numbers or characters}

\item{min_tau}{numeric indicating the smallest gap between subsequent numbers for which
the autocorrelation will be computed}

\item{max_tau}{numeric indicating the largest gap between subsequent numbers for which
the autocorrelation will be computed}
}
\value{
maximum autocorrelation of \code{x} for the all gaps \code{tau} in
the interval from \code{min_tau} to \code{max_tau} between subsequent numbers
}
\description{
Compute the Maximum Autocorrelation
}
\details{
This function takes a vector \code{x} and uses it to compute the maximum autocorrelation
for a specified interval of gaps tau from \code{min_tau} to \code{max_tau}.
The resulting autocorrelation is the largest autocorrelation that could be observed
in the interval for values of \code{tau}. A high autocorrelation
indicates that there there is a high correlation between numbers in a
sequence that are set apart from each other \code{min_tau} to \code{max_tau} values.
}
\examples{
max_auto_corr(c(1, 1, 1, 2, 2, 2, 1, 2, 1, 2, 2, 2), tau = 2)
max_auto_corr(c(1, 1, 1, 2, 2, 2, 1, 2, 1, 2, 2, 2), tau = 3)

}
