% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{null_score}
\alias{null_score}
\title{Guttmann's Null-Score Quotient}
\usage{
null_score(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
Guttmann Null-Score Quotient of \code{x}
}
\description{
Guttmann's Null-Score Quotient
}
\details{
This function takes a vector \code{x} and computes the number of all response
pairs that do not occur in this sequence given the number of possible
\code{options}. The result of this computation is weighted with the highest
possible number of unused response pairs for this sequence.
The final result is attained by multiplying this quotient with 100.
The result of this quotient ranges from 0 to 100.
High values indicate an uneven distribution of response pairs, whereas low
values indicate an even distribution of response pairs.
}
\references{
Towse, J.N., Neil, D. Analyzing human random generation behavior: A review of
methods used and a computer program for describing performance. Behavior
Research Methods, Instruments, & Computers 30, 583–591 (1998).
\url{https://doi.org/10.3758/BF03209475}
}
