% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_index.R
\name{phi_index}
\alias{phi_index}
\title{Phi Index}
\usage{
phi_index(x, options, order = 2)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}

\item{order}{order of analysis}
}
\value{
Phi index of \code{x}
}
\description{
Phi Index
}
\details{
to be filled
}
\references{
Towse, J.N., Neil, D. Analyzing human random generation behavior: A review of
methods used and a computer program for describing performance. Behavior
Research Methods, Instruments, & Computers 30, 583–591 (1998).
\url{https://doi.org/10.3758/BF03209475}

Wagenaar, W. A. 1970. “Subjective Randomness and the Capacity to
Generate Information.” Acta Psychologica 33:233–42.
doi: 10.1016/0001-6918(70)90135-6.
}
