% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy.R
\name{redundancy_index}
\alias{redundancy_index}
\title{Compute redundancy index (see Towse & Neil, 1998)}
\usage{
redundancy_index(x, possible_responses)
}
\arguments{
\item{x}{vector of random numbers}

\item{possible_responses}{number of available options in sequence}
}
\value{
redundancy index of \code{x}
}
\description{
Compute redundancy index (see Towse & Neil, 1998)
}
