% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_order_regularity.R
\name{reg_index}
\alias{reg_index}
\title{Compute regularity index}
\usage{
reg_index(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
regularity index of \code{x}
}
\description{
Compute regularity index
}
\references{
Skliar, Monge & Oviedo (2009)
}
