% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_indices.R
\name{rng2_index}
\alias{rng2_index}
\title{Compute RNG2 index}
\usage{
rng2_index(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
RNG2 index of \code{x}
}
\description{
Compute RNG2 index
}
\details{
This function takes a vector \code{x} and computes an index that reflects
whether the distribution of response pairs in the sequence is uneven given
the number of possible \code{options}. In this modified version of the
RNG index response pairs are constructed over the distance of two values,
e.g., in the sequence c(1,2,3) the first response pair is (1,3).
The values of this index range between 0 and 1. A value of '1' would indicate
complete redundancy of response pairs,whereas a value of '0' would indicate
an even distribution of response pairs.
}
\references{
Towse, J.N., Neil, D. Analyzing human random generation behavior: A review of
methods used and a computer program for describing performance. Behavior
Research Methods, Instruments, & Computers 30, 583–591 (1998).
}
