% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{rng_index}
\alias{rng_index}
\title{Compute RNG index by Evans (1978)}
\usage{
rng_index(x, possible_responses)
}
\arguments{
\item{x}{vector of random numbers}

\item{possible_responses}{number of available options in sequence}
}
\value{
RNG index of \code{x}
}
\description{
Compute RNG index by Evans (1978)
}
