% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularity_indices.R
\name{runs_index}
\alias{runs_index}
\title{Compute runs (see Ginsburg & Karpiuk, 1994)}
\usage{
runs_index(x, asc = TRUE)
}
\arguments{
\item{x}{vector of random numbers}
}
\value{
runs of \code{x}
}
\description{
Compute the variance of successive ascending or descending runs
}
