% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{series}
\alias{series}
\title{Compute series as measure of randomness (see Gisburg & Karpiuk, 1994)}
\usage{
series(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
series of \code{x}
}
\description{
Compute frequency with which values are followed by their most
adjacent predecessors and successors in the vector
}
