% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy.R
\name{to_numeric}
\alias{to_numeric}
\title{function transforms given vector to numeric}
\usage{
to_numeric(x)
}
\arguments{
\item{x}{vector}
}
\value{
x_transformed
}
\description{
function transforms given vector to numeric
}
\keyword{internal}
