% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turning_point_index.R
\name{tp_index}
\alias{tp_index}
\title{Turning Point Index}
\usage{
tp_index(x)
}
\arguments{
\item{x}{vector of random numbers}
}
\value{
Turning Point Index of \code{x}
}
\description{
Turning Point Index
}
\details{
This function takes a vector \code{x} and computes all turning points in the
given sequence. Turning points are defined as values in the given
sequence where the preceding and succeeding value are both either higher
or lower than the current value. The number of observed turning points is
weighted against the theoretically expected number of turning points given
the length of the sequence. The resulting quotient is multiplied with 100.
Values over 100 indicate a higher number of turning points than theoretically
expected; values below 100 indicate fewer turning points than theoretically
expected.
}
\examples{
tp_index(ginsburg1994)
tp_index(evans1978[, 1])
tp_index(evans1978[, 2])

}
\references{
Towse, J.N., Neil, D. Analyzing human random generation behavior: A review of
methods used and a computer program for describing performance. Behavior
Research Methods, Instruments, & Computers 30, 583–591 (1998).
\doi{doi:10.3758/BF03209475}
}
