% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_order_dependencies.R
\name{var_digits}
\alias{var_digits}
\title{Variance of digits}
\usage{
var_digits(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
variance of digits of \code{x}
}
\description{
Computes the variance of marginal totals
}
\details{
This function takes a vector \code{x} and computes the population variance
of response frequencies given the number of possible \code{options}.
This index is conceptually closely related to the idea of the
\code{\link{redundancy_index}}.
}
\examples{
var_digits(ginsburg1994, 10)
var_digits(evans1978[, 1], 10)
var_digits(evans1978[, 2], 10)

}
\references{
Ginsburg N, Karpiuk P. Random Generation: Analysis of the Responses.
Perceptual and Motor Skills. 1994;79(3):1059-1067.
\url{doi:10.2466/pms.1994.79.3.1059}
}
