% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_order_dependencies.R
\name{variance_of_digits}
\alias{variance_of_digits}
\title{Compute variance of digits (see Ginsburg & Karpiuk, 1994)}
\usage{
variance_of_digits(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
variance of digits of \code{x}
}
\description{
Computes the variance of marginal totals
}
