% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_indices.R
\name{all_rand}
\alias{all_rand}
\title{Compute variety of randomness indices}
\usage{
all_rand(df, columns = NULL, indices_names = NULL, new_col_name)
}
\arguments{
\item{df}{data frame containing sequences of options in row-wise format}

\item{columns}{columns which will be interpreted as sequence}

\item{indices_names}{indices of randomness to be computed}
}
\description{
Compute variety of randomness indices
}
\details{
This function allows to enter a data frame of sequence data and computes a
variety of randomness indices. It is assumed that the data is provided in
row-wise, i.e., each row represents one sequences. The output of this function
is also a data frame.
}
