% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{cluster_ratio}
\alias{cluster_ratio}
\title{Compute cluster ratio as a measure of randomness (see Ginsburg & Karpiuk, 1994)}
\usage{
cluster_ratio(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
cluster ratio of \code{x}
}
\description{
compute the variance of all frequencies in table of
first-order dependencies
}
\details{
This function takes a vector \code{x} and computes the population variance
of all response pair frequencies given the number of possible \code{options}.
}
\references{
Ginsburg N, Karpiuk P. Random Generation: Analysis of the Responses.
Perceptual and Motor Skills. 1994;79(3):1059-1067.
\url{doi:10.2466/pms.1994.79.3.1059}
}
