% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{cluster_ratio}
\alias{cluster_ratio}
\title{Compute cluster ratio as a measure of randomness (see Ginsburg & Karpiuk, 1994)}
\usage{
cluster_ratio(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
cluster ratio of \code{x}
}
\description{
compute the variance of all frequencies in table of
first-order dependencies
}
