% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{cluster_ratio}
\alias{cluster_ratio}
\title{Cluster ratio index}
\usage{
cluster_ratio(x, options)
}
\arguments{
\item{x}{vector of distinct options of numbers or characters}

\item{options}{number of available options in sequence}
}
\value{
cluster ratio of \code{x}
}
\description{
compute the population variance of all response pair frequencies
}
\details{
This function takes a vector \code{x} and computes the population variance
of all response pair frequencies given the number of possible \code{options}.
}
\examples{
cluster_ratio(ginsburg1994, 10)
cluster_ratio(evans1978[, 1], 10)
cluster_ratio(evans1978[, 2], 10)

}
\references{
Ginsburg N, Karpiuk P. Random Generation: Analysis of the Responses.
Perceptual and Motor Skills. 1994;79(3):1059-1067.
\doi{doi:10.2466/pms.1994.79.3.1059}
}
