% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{convert_to_matrix}
\alias{convert_to_matrix}
\title{convert vector to matrix of first order dependencies}
\usage{
convert_to_matrix(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\description{
helper function to transform a given vector to a matrix of first
order dependencies, specifying the frequency of all possible 2-digit long
sequences
}
\keyword{internal}
