% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularity_indices.R
\name{coupon_score}
\alias{coupon_score}
\title{Coupon Score}
\usage{
coupon_score(x, options)
}
\arguments{
\item{x}{vector of distinct options of numbers or characters}

\item{options}{number of available options in sequence}
}
\value{
Coupon Score of \code{x}
}
\description{
Compute average digit length required for all responses to occur
}
\details{
This function takes a vector \code{x} and computes the mean number of values
that is required for all possible \code{options} to occur. This means that the algorithm
starts to count at the beginning of the sequence until all \code{options} have occurred.
Having observed all \code{options},
the counting process starts again at zero at the next value of the sequence.
Finally, the mean over the length of these
complete sets of values (containing all options) is computed.
Incomplete sets of responses at the end of a sequence with at least one
option missing are not used for the computation of this index.
Consequently, this index cannot be computed for vectors that do not
contain all possible \code{options} and therefore, NA is returned. The lowest
possible score equals the number of \code{options}, indicating that
each complete set contains every possible option only once. A higher score
indicates that at least one complete set contains values of the same kind at least twice.
}
\examples{
runs_index(ginsburg1994, 10)
runs_index(evans1978[, 1], 10)
runs_index(evans1978[, 2], 10)

}
\references{
Ginsburg N, Karpiuk P. Random Generation: Analysis of the Responses.
Perceptual and Motor Skills. 1994;79(3):1059-1067.
\doi{doi:10.2466/pms.1994.79.3.1059}

Towse, J.N., Neil, D. Analyzing human random generation behavior: A review of
methods used and a computer program for describing performance. Behavior
Research Methods, Instruments, & Computers 30, 583–591 (1998).
\doi{doi:10.3758/BF03209475}
}
