% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{digram_rep}
\alias{digram_rep}
\title{Compute Digram Repetitions (see Ginsburg & Karpiuk, 1994)}
\usage{
digram_rep(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
digram repetition of \code{x}
}
\description{
Compute Digram Repetitions (see Ginsburg & Karpiuk, 1994)
}
