% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{digram_rep}
\alias{digram_rep}
\title{Digram Repetitions}
\usage{
digram_rep(x, options)
}
\arguments{
\item{x}{vector of distinct options of numbers or characters}

\item{options}{number of available options in sequence}
}
\value{
digram repetition of \code{x}
}
\description{
Compute sum of all response pairs minus one
}
\details{
This function takes a vector \code{x} and sums up the frequency of all
response pairs given the number of possible \code{options}. For each response
pair that occurs at least once the total sum is reduced by one. A lower score
indicates an even distribution of response pairs, i.e., most or all of them occurred.
A higher score indicates an uneven distribution of response pairs, i.e., many
response pairs did not occur.
}
\examples{
digram_rep(ginsburg1994, 10)
digram_rep(evans1978[, 1], 10)
digram_rep(evans1978[, 2], 10)

}
\references{
Ginsburg N, Karpiuk P. Random Generation: Analysis of the Responses.
Perceptual and Motor Skills. 1994;79(3):1059-1067.
\doi{doi:10.2466/pms.1994.79.3.1059}
}
