% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularity_indices.R
\name{gap_score}
\alias{gap_score}
\title{Gap Score}
\usage{
gap_score(x)
}
\arguments{
\item{x}{vector of distinct options of numbers or characters}
}
\value{
gap score of \code{x}
}
\description{
Compute median gap between identical values
}
\details{
This function takes a vector \code{x} and computes the median gap between
the most adjacent identical values.
}
\examples{
gap_score(ginsburg1994)
gap_score(evans1978[, 1])
gap_score(evans1978[, 2])

}
\references{
#' Ginsburg N, Karpiuk P. Random Generation: Analysis of the Responses.
Perceptual and Motor Skills. 1994;79(3):1059-1067.
\doi{doi:10.2466/pms.1994.79.3.1059}
}
