% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularity_indices.R
\name{gap_score}
\alias{gap_score}
\title{Compute gap score (see Ginsburg & Karpiuk, 1994)}
\usage{
gap_score(x)
}
\arguments{
\item{x}{vector of random numbers}
}
\value{
gap score of \code{x}
}
\description{
Compute median gap between identical values
}
