% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularity_indices.R
\name{gap_score}
\alias{gap_score}
\title{Gap Score}
\usage{
gap_score(x)
}
\arguments{
\item{x}{vector of distinct options of numbers or characters}
}
\value{
gap score of \code{x}
}
\description{
Compute median gap between identical values
}
\details{
This function takes a vector \code{x} and computes the median gap between
the most adjacent identical values. For each value encountered in the sequence
it is computed whether the same value occurs again in the sequence. If so,
the length of the gap between them is stored. Finally, the mean over these
gap lengths is computed and returned.
The lowest score of 1 indicates that most or all values of the same kind are
right next to each other. A higher scores indicates that options of the
same kind often have a gap between them containing options of a different kind.
}
\examples{
gap_score(ginsburg1994)
gap_score(evans1978[, 1])
gap_score(evans1978[, 2])

}
\references{
#' Ginsburg N, Karpiuk P. Random Generation: Analysis of the Responses.
Perceptual and Motor Skills. 1994;79(3):1059-1067.
\doi{doi:10.2466/pms.1994.79.3.1059}
}
