% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularity_indices.R
\name{poker_score}
\alias{poker_score}
\title{Compute Poker Score (see Ginsburg & Karpiuk, 1994)}
\usage{
poker_score(x)
}
\arguments{
\item{x}{vector of random numbers}
}
\value{
poker score of \code{x}
}
\description{
Compute number of times exactly two responses of the same value
occur after division of the complete sequence into 5-digit-long sub-sequences
}
