% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_order_dependencies.R
\name{redundancy_index}
\alias{redundancy_index}
\title{Redundancy index}
\usage{
redundancy_index(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
redundancy index of \code{x}
}
\description{
Computes index indicating the evenness of the response frequency
distribution.
}
\details{
This function takes a vector \code{x} and weighs the observed information
concerning the response frequencies against the highest observable
information given the number of possible \code{options}. The resulting
quotient is subtracted from 1 and then, multiplied with 100.
A value of 0 would indicate an even distribution of response frequencies,
whereas a value of 100 would indicate complete redundancy of one response.
}
\examples{
redundancy_index(c(1,1,1,1),2)
redundancy_index(c(2,2,1,1),2)

}
\references{
Towse, J.N., Neil, D. Analyzing human random generation behavior: A review of
methods used and a computer program for describing performance. Behavior
Research Methods, Instruments, & Computers 30, 583–591 (1998).
\url{https://doi.org/10.3758/BF03209475}
}
