% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_order_dependencies.R
\name{redundancy_index}
\alias{redundancy_index}
\title{Compute redundancy index (see Towse & Neil, 1998)}
\usage{
redundancy_index(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
redundancy index of \code{x}
}
\description{
Compute redundancy index (see Towse & Neil, 1998)
}
\examples{
redundancy_index(c(1,1,1,1),2)
redundancy_index(c(2,2,1,1),2)

}
