% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_order_regularity.R
\name{reg_index}
\alias{reg_index}
\title{Compute regularity index}
\usage{
reg_index(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
regularity index of \code{x}
}
\description{
Compute regularity index
}
\details{
This function takes a vector \code{x} as a sequence and the number of
possible \code{options} and computes the regularity of the given sequence.
Values range from 0 and 1. A value of 0 (which is barely ever obtained)
indicates no regularity, whereas a value of 1 indicates maximum regularity.
The regularity index is obtained by identifying the the maximum regularity
occurring in the sequence.
}
\examples{
reg_index(ginsburg1994, 10)
reg_index(evans1978[, 1], 10)
reg_index(evans1978[, 2], 10)

}
\references{
Skliar, Osvaldo, Ricardo E. Monge, Guillermo Oviedo, and Víctor Medina. 2009.
“Indices of regularity and indices of randomness for m-ary strings.”
\url{doi:10.15517/rmta.v16i1.1418}.
}
