% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{repetitions}
\alias{repetitions}
\title{Repetitions index of randomness}
\usage{
repetitions(x)
}
\arguments{
\item{x}{vector of random numbers}
}
\value{
repetitions of \code{x}
}
\description{
Compute the frequency a value is repeated in the next round
}
\details{
This function takes a vector \code{x} and counts how often values are
followed by the same value in the next position of the sequence.
}
\examples{
repetitions(ginsburg1994)
repetitions(evans1978[, 1])
repetitions(evans1978[, 2])

}
\references{
Ginsburg N, Karpiuk P. Random Generation: Analysis of the Responses.
Perceptual and Motor Skills. 1994;79(3):1059-1067.
\doi{doi:10.2466/pms.1994.79.3.1059}
}
