% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{repetitions}
\alias{repetitions}
\title{Compute repetitions as measure of randomness (see Ginsburg & Karpiuk, 1994)}
\usage{
repetitions(x, options)
}
\arguments{
\item{x}{vector of random numbers}

\item{options}{number of available options in sequence}
}
\value{
repetitions of \code{x}
}
\description{
Compute frequency a value is repeated in the next round
}
