% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_indices.R
\name{rng_index}
\alias{rng_index}
\title{Compute RNG index}
\usage{
rng_index(x, options, circ = TRUE)
}
\arguments{
\item{x}{vector of distinct options of numbers or characters}

\item{options}{number of available options in sequence}

\item{circ}{indicate whether to include wrap around from end to the beginning
of the sequence when computing response pairs}
}
\value{
RNG index of \code{x}
}
\description{
Compute RNG index
}
\details{
This function takes a vector \code{x} and computes an index that reflects
whether the distribution of response pairs in the sequence is uneven given
the number of possible \code{options}. The scores of this index range between
0 and 1. A score of 1 indicates complete redundancy of one response pair (e.g., '1-2'),
whereas a score of 0 indicates an even distribution of response pairs.
}
\examples{
rng_index(ginsburg1994, 10)
rng_index(ginsburg1994, 10, circ = FALSE)
rng_index(evans1978[, 1], 10)
rng_index(evans1978[, 2], 10)

}
\references{
Evans, F.J. Monitoring attention deployment by random number generation:
An index to measure subjective randomness.
Bull. Psychon. Soc. 12, 35–38 (1978).
\doi{doi:10.3758/BF03209475}
}
