% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularity_indices.R
\name{runs_index}
\alias{runs_index}
\title{Compute runs index of randomness}
\usage{
runs_index(x, asc = TRUE)
}
\arguments{
\item{x}{vector of distinct options of numbers or characters}

\item{asc}{Indicate whether to compute variance of ascending or descending
runs. Default value is set to ascending.}
}
\value{
runs of \code{x}
}
\description{
Compute runs index of randomness
}
\details{
This function takes a vector \code{x} and computes the population variance
of all ascending or descending run lengths, depending on \code{asc}.
An ascending run is defined by a value in a sequence being followed by
a larger value in the next position of the sequence.
The default version of this function computes the population variance of
ascending run lengths.
}
\examples{
runs_index(ginsburg1994)
runs_index(ginsburg1994, FALSE)
runs_index(evans1978[, 1])
runs_index(evans1978[, 2])

}
\references{
Ginsburg N, Karpiuk P. Random Generation: Analysis of the Responses.
Perceptual and Motor Skills. 1994;79(3):1059-1067.
\doi{doi:10.2466/pms.1994.79.3.1059}
}
