% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_order_dependencies.R
\name{series}
\alias{series}
\title{Series index of randomness}
\usage{
series(x, options)
}
\arguments{
\item{x}{vector of distinct options of numbers or characters}

\item{options}{number of available options in sequence}
}
\value{
series of \code{x}
}
\description{
Compute frequency with which values are followed by their most
adjacent predecessors and successors
}
\details{
This function takes a vector \code{x} and counts how often values are
followed by their most adjacent predecessor and successors given the number
of possible \code{options}. This function includes series from the lowest to
the highest value and vice versa.
An example of a series would be '1,2,3' (options = 3).
The resulting score of this sequence is 3 (from 1 to 2, from 2 to 3,
and from 3 to 1).
The highest possible score equals the number of values in a sequence when
each value is followed by its most adjacent predecessor or successor. The
lowest possible score is 0 when each value is followed by another value than
its predecessor or successor.
}
\examples{
series(ginsburg1994, 10)
series(evans1978[, 1], 10)
series(evans1978[, 2], 10)

}
\references{
Ginsburg N, Karpiuk P. Random Generation: Analysis of the Responses.
Perceptual and Motor Skills. 1994;79(3):1059-1067.
\doi{doi:10.2466/pms.1994.79.3.1059}
}
